# docs/swagger.yaml
basePath: /
definitions:
  main.ProductDTO:
    properties:
      code:
        type: string
      price:
        type: integer
    type: object
  main.UserDTO:
    properties:
      email:
        type: string
      password:
        type: string
      verified_email:
        type: boolean
    type: object
host: localhost:9000
info:
  contact:
    email: marckenneth.lomio@gmail.com
    name: Marc Kenneth Lomio
  description: Go Rest API with SQL SERVER DB
  title: Go Rest API
  version: "1.0"
paths:
  /:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: html Home Page that have link for google log-in Oauth2
  /googlecallback:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Google send user data
      tags:
      - 3rdParty
  /migration:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Migrate tables to the SQL Server
      tags:
      - migration
  /product:
    post:
      parameters:
      - description: Create product dto
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/main.ProductDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ProductDTO'
      summary: Create Product
      tags:
      - products
  /product/{id}:
    delete:
      parameters:
      - description: Delete product by id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Delete specific Product
      tags:
      - products
    get:
      parameters:
      - description: Get product by id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ProductDTO'
      summary: Get specific Product
      tags:
      - products
    put:
      parameters:
      - description: Get product by id
        in: path
        name: id
        required: true
        type: string
      - description: Update product dto
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/main.ProductDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ProductDTO'
      summary: Update Product
      tags:
      - products
  /products:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Get all products
      tags:
      - products
  /user:
    post:
      parameters:
      - description: Create user dto
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.UserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.UserDTO'
      summary: Create User
      tags:
      - users
  /user/{authcode}:
    get:
      parameters:
      - description: Get user by authcode
        in: path
        name: authcode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.UserDTO'
      summary: Get user by authcode
      tags:
      - users
  /user/login:
    post:
      parameters:
      - description: Login user dto
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.UserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.UserDTO'
      summary: Login User
      tags:
      - users
  /user/loginViaGoogle:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Login user via google account
      tags:
      - users
  /user/verification/{id}:
    get:
      parameters:
      - description: Get user by id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.UserDTO'
      summary: Verify user
      tags:
      - users
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Get all users
      tags:
      - users
swagger: "2.0"